requirements.txt
python-telegram-bot==13.15
requests==2.31.0
pandas==2.1.1
python-dotenv==1.0.0
urllib3==1.26.18
APScheduler==3.6.3
cachetools==4.2.2
certifi>=2023.7.22
pytz==2025.2
tornado==6.1
tzlocal==5.3.1
six==1.17.0
setuptools>=65.5.0
